adds model.go file and thus takes all models out of the commands. Attention Nef10, this also deletes the helper stuff because it would not work otherwise.
