+++ commands/builds.go
@@ -24,30 +24,6 @@ func init() {
 	)
 }
 
-type Builds struct {
-	Builds []Build `json:"builds"`
-}
-
-type Build struct {
-	Number     string  `json:"number"`
-	State      string  `json:"state"`
-	StartedAt  string  `json:"started_at"`
-	FinishedAt string  `json:"finished_at"`
-	Duration   int     `json:"duration"`
-	EventType  string  `json:"event_type"`
-	Branch     *Branch `json:"branch"`
-	Commit     *Commit `json:"commit"`
-	Jobs       Jobs    `json:"jobs"`
-}
-
-func (b *Build) HasPassed() bool {
-	return b.State == "passed"
-}
-
-func (b *Build) IsNotYetFinished() bool {
-	return ((b.State == "created") || (b.State == "started"))
-}
-
 func PushColorAccordingToBuildStatusBold(build Build, cmd *cli.Cmd) {
 	if build.HasPassed() {
 		cmd.Stdout.PushColor("boldgreen")
@@ -70,19 +46,6 @@ func PushColorAccordingToBuildStatus(build Build, cmd *cli.Cmd) {
 			cmd.Stdout.PushColor("red")
 		}
 	}
-}
-
-type Commit struct {
-	Message string `json:"message"`
-}
-
-type Jobs struct {
-	Jobs []Job `json:"jobs"`
-}
-
-type Job struct {
-	Number string `json:"number"`
-	State  string `json:"state"`
 }
 
 func buildsCmd(cmd *cli.Cmd) {
