+++ commands/whoami.go
@@ -1,7 +1,6 @@
 package commands
 
 import (
-	"github.com/HPI-BP2015H/go-travis/commands/helper"
 	"github.com/HPI-BP2015H/go-travis/config"
 	"github.com/HPI-BP2015H/go-utils/cli"
 )
@@ -19,7 +18,7 @@ func init() {
 func whoamiCmd(cmd *cli.Cmd) {
 	env := cmd.Env.(config.TravisCommandConfig)
 
-	user, err := user.CurrentUser(env.Client)
+	user, err := CurrentUser(env.Client)
 	if err != nil {
 		cmd.Stderr.Println("Error: Could not get the current user!")
 		cmd.Exit(1)
@@ -28,7 +27,7 @@ func whoamiCmd(cmd *cli.Cmd) {
 	cmd.Exit(0)
 }
 
-func printUser(user user.User, cmd *cli.Cmd) {
+func printUser(user User, cmd *cli.Cmd) {
 	cmd.Stdout.PushColor("green")
 	cmd.Stdout.Print("You are ")
 	cmd.Stdout.Cprint("boldgreen", user.Login)
